feat(megamenu): add MegaMenuBox and rewrite megamenu
