+++ src/components/MegaMenu/MegaMenuBox/MegaMenuBox.css
@@ -1,0 +1,66 @@
+.che--MegaMenuBox {
+  &-Wrapper {
+    width: 100%;
+    width: var(--mega-menu-content-width);
+    max-width: var(--mega-menu-content-width);
+    pointer-events: all;
+  }
+
+  &-Window {
+    position: absolute;
+    z-index: 2;
+    top: var(--mega-menu-offset);
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    width: 100%;
+    height: max-content;
+    background: var(--color-bg-default);
+    pointer-events: none;
+    transition: transform 0.24s ease, opacity 0.4s;
+
+    &_animate {
+      &_exited,
+      &_exiting {
+        opacity: 0;
+        transform: var(--mega-menu-transform);
+      }
+
+      &_entered,
+      &_entering {
+        opacity: 1;
+        transform: translate3d(0, 0, 0);
+      }
+    }
+  }
+
+  &-Overlay {
+    position: fixed;
+    top: 0;
+    right: 0;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    height: 100%;
+    background: var(--color-bg-tone);
+
+    &_animate {
+      &_exited,
+      &_exiting {
+        opacity: 0;
+      }
+
+      &_entered,
+      &_entering {
+        opacity: 1;
+      }
+
+      &_exited,
+      &_entered,
+      &_entering,
+      &_exiting {
+        transition: opacity 0.24s ease;
+      }
+    }
+  }
+}
