+++ src/components/MegaMenu/MegaMenuBox/index.ts
@@ -1,0 +1,2 @@
+export * from './MegaMenuBox';
+export * from './types';
