+++ src/components/MegaMenu/__stand__/MegaMenu.variants.css
@@ -1,5 +1,8 @@
 .che--MegaMenuVariants {
-  display: flex;
+  /* display: flex;
   justify-content: flex-start;
   align-items: flex-start;
+  height: 100%;
+  width: 100%;
+  overflow: auto; */
 }
