+++ src/components/MegaMenu/index.ts
@@ -1,0 +1,3 @@
+export * from './MegaMenu';
+export * from './types';
+export * from './MegaMenuBox';
