--- src/components/MegaMenu/index.tsx
@@ -1,1 +1,0 @@
-export * from './MegaMenu';
